refactor: changing moduleResolution to bundler
