--- apps/backend/parsers/formula-parser/TODO.md
@@ -1,4 +1,0 @@
-# TODOs
-
-- migrate to typescript
-- serde (deserialise/serialize)
