+++ apps/backend/parsers/formula-parser/src/core/index.ts
@@ -1,2 +1,2 @@
-export * from "./config.ts";
-export * from "./errors.ts";
+export * from "./config";
+export * from "./errors";
