+++ apps/backend/parsers/formula-parser/src/utils/index.ts
@@ -1,3 +1,3 @@
-export * from "./logger.ts";
-export * from "./proto-utils.ts";
-export * from "./to-proto.ts";
+export * from "./logger";
+export * from "./proto-utils";
+export * from "./to-proto";
