+++ apps/backend/parsers/formula-parser/tsconfig.json
@@ -2,8 +2,8 @@
     "extends": "@etl-design/tools-tsconfig/tsconfig.json",
     "compilerOptions": {
         "target": "es2022",
-        "module": "nodenext",
-        "moduleResolution": "node16",
+        "module": "esnext",
+        "moduleResolution": "bundler",
         "paths": {
             "@/*": ["./src/*"],
             "@@/*": ["./*"]
